Split out utils
