+++ src/Distribution/Server/Features/TarIndexCache.hs
@@ -20,6 +20,7 @@ import Distribution.Server.Framework.BackupRestore
 import qualified Distribution.Server.Features.TarIndexCache.State as Acid
 import Distribution.Server.Features.Users
 import Distribution.Server.Packages.Types
+import Distribution.Server.Packages.Utils
 import Data.TarIndex
 import qualified Data.TarIndex as TarIndex
 import Distribution.Server.Util.ServeTarball (constructTarIndex)
