+++ src/Distribution/Server/Packages/Index.hs
@@ -18,7 +18,9 @@ import Distribution.Server.Framework.MemSize
 import Distribution.Server.Packages.Types
          ( CabalFileText(..), PkgInfo(..)
          , TarballRevIx, MetadataRevIx
-         , pkgSpecificRevision
+         )
+import Distribution.Server.Packages.Utils
+         ( pkgSpecificRevision
          , pkgLatestCabalFileText, pkgLatestUploadInfo
          )
 import Distribution.Server.Packages.Metadata
@@ -41,7 +43,6 @@ import Data.SafeCopy (base, deriveSafeCopy)
 
 import Data.Map (Map)
 import qualified Data.Map as Map
-import qualified Data.Vector as Vec
 import Data.ByteString.Lazy (LazyByteString, fromStrict)
 import System.FilePath.Posix
 import Data.Maybe (mapMaybe)
